+++ playtime-0.10.0/src/api/jwt-utils.js
@@ -5,10 +5,15 @@ import dotenv from "dotenv";
 const result = dotenv.config();
 
 export function createToken(user) {
-  return jwt.sign({ id: user._id, email: user.email }, "secretpasswordnotrevealedtoanyone", {
+  const payload = {
+    id: user._id,
+    email: user.email
+  };
+  const options = {
     algorithm: "HS256",
     expiresIn: "1h"
-  });
+  };
+  return jwt.sign(payload, process.env.cookie_password, options);
 }
 
 export function decodeToken(token) {
@@ -18,8 +23,8 @@ export function decodeToken(token) {
     userInfo.userId = decoded.id;
     userInfo.email = decoded.email;
   } catch (e) {
+    console.log(e.message);
   }
-
   return userInfo;
 }
 
@@ -28,19 +33,6 @@ export async function validate(decoded, request) {
   if (!user) {
     return { isValid: false };
   } else {
-    return { isValid: true };
+    return { isValid: true, credentials: user };
   }
-}
-
-export function getUserIdFromRequest(request) {
-  var userId = null;
-  try {
-    const authorization = request.headers.authorization;
-    var token = authorization.split(" ")[1];
-    var decodedToken = jwt.verify(token, "secretpasswordnotrevealedtoanyone");
-    userId = decodedToken.id;
-  } catch (e) {
-    userId = null;
-  }
-  return userId;
 }