+++ playtime-0.10.0/src/api/playlist-api.js
@@ -5,7 +5,9 @@ import { validationError } from "./logger.js";
 
 export const playlistApi = {
   find: {
-    auth: false,
+    auth: {
+      strategy: "jwt",
+    },
     handler: async function (request, h) {
       try {
         const playlists = await db.playlistStore.getAllPlaylists();
@@ -21,7 +23,9 @@ export const playlistApi = {
   },
 
   findOne: {
-    auth: false,
+    auth: {
+      strategy: "jwt",
+    },
     async handler(request) {
       try {
         const playlist = await db.playlistStore.getPlaylistById(request.params.id);
@@ -41,7 +45,9 @@ export const playlistApi = {
   },
 
   create: {
-    auth: false,
+    auth: {
+      strategy: "jwt",
+    },
     handler: async function (request, h) {
       try {
         const playlist = request.payload;
@@ -62,7 +68,9 @@ export const playlistApi = {
   },
 
   deleteOne: {
-    auth: false,
+    auth: {
+      strategy: "jwt",
+    },
     handler: async function (request, h) {
       try {
         const playlist = await db.playlistStore.getPlaylistById(request.params.id);
@@ -81,7 +89,9 @@ export const playlistApi = {
   },
 
   deleteAll: {
-    auth: false,
+    auth: {
+      strategy: "jwt",
+    },
     handler: async function (request, h) {
       try {
         await db.playlistStore.deleteAllPlaylists();