+++ playtime-0.10.0/src/api/user-api.js
@@ -1,13 +1,15 @@
 import Boom from "@hapi/boom";
 import { db } from "../models/db.js";
+import { IdSpec, JwtAuth, UserArray, UserSpec, UserCredentialsSpec, UserSpecPlus } from "../models/joi-schemas.js";
 import { validationError } from "./logger.js";
-import { IdSpec, UserArray, UserSpec, UserSpecPlus } from "../models/joi-schemas.js";
 import { createToken } from "./jwt-utils.js";
 
 export const userApi = {
   find: {
-    auth: false,
-    handler: async function (request, h) {
+    auth: {
+      strategy: "jwt"
+    },
+    handler: async function(request, h) {
       try {
         const users = await db.userStore.getAllUsers();
         return users;
@@ -18,12 +20,14 @@ export const userApi = {
     tags: ["api"],
     description: "Get all userApi",
     notes: "Returns details of all userApi",
-    response: { schema: UserArray, failAction: validationError },
+    response: { schema: UserArray, failAction: validationError }
   },
 
   findOne: {
-    auth: false,
-    handler: async function (request, h) {
+    auth: {
+      strategy: "jwt"
+    },
+    handler: async function(request, h) {
       try {
         const user = await db.userStore.getUserById(request.params.id);
         if (!user) {
@@ -38,12 +42,12 @@ export const userApi = {
     description: "Get a specific user",
     notes: "Returns user details",
     validate: { params: { id: IdSpec }, failAction: validationError },
-    response: { schema: UserSpecPlus, failAction: validationError },
+    response: { schema: UserSpecPlus, failAction: validationError }
   },
 
   create: {
     auth: false,
-    handler: async function (request, h) {
+    handler: async function(request, h) {
       try {
         const user = await db.userStore.addUser(request.payload);
         if (user) {
@@ -58,12 +62,14 @@ export const userApi = {
     description: "Create a User",
     notes: "Returns the newly created user",
     validate: { payload: UserSpec, failAction: validationError },
-    response: { schema: UserSpecPlus, failAction: validationError },
+    response: { schema: UserSpecPlus, failAction: validationError }
   },
 
   deleteAll: {
-    auth: false,
-    handler: async function (request, h) {
+    auth: {
+      strategy: "jwt"
+    },
+    handler: async function(request, h) {
       try {
         await db.userStore.deleteAll();
         return h.response().code(204);
@@ -73,29 +79,30 @@ export const userApi = {
     },
     tags: ["api"],
     description: "Delete all userApi",
-    notes: "All users removed from Playtime",
+    notes: "All userApi removed from Playtime"
   },
 
   authenticate: {
     auth: false,
-    handler: async function (request, h) {
+    handler: async function(request, h) {
       try {
         const user = await db.userStore.getUserByEmail(request.payload.email);
         if (!user) {
           return Boom.unauthorized("User not found");
-        }
-        if (user.password !== request.payload.password) {
+        } else if (user.password !== request.payload.password) {
           return Boom.unauthorized("Invalid password");
+        } else {
+          const token = createToken(user);
+          return h.response({ success: true, token: token }).code(201);
         }
-        const token = createToken(user);
-        return h.response({ success: true, token: token }).code(201);
       } catch (err) {
         return Boom.serverUnavailable("Database Error");
       }
     },
     tags: ["api"],
-    description: "Authenticate",
-    notes: "Authenticate a user and retrieve a JWT token",
-    // validate: { payload: UserCredentialsSpec, failAction: validationError },
-  },
+    description: "Authenticate  a User",
+    notes: "All userApi removed from Playtime",
+    validate: { payload: UserCredentialsSpec, failAction: validationError },
+    response: { schema: JwtAuth, failAction: validationError }
+  }
 };