+++ playtime-0.10.0/src/models/joi-schemas.js
@@ -51,3 +51,10 @@ export const PlaylistSpecPlus = PlaylistSpec.keys({
 }).label("PlaylistPlus");
 
 export const PlaylistArraySpec = Joi.array().items(PlaylistSpecPlus).label("PlaylistArray");
+
+export const JwtAuth = Joi.object()
+  .keys({
+    success: Joi.boolean().example("true").required(),
+    token: Joi.string().example("eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjYyMWNkNjhlMzU4NTA1NmVmZDg5YmJiYyIsImVtYWlsIjoiaG9tZXJAc2ltcHNvbi5jb20iLCJpYXQiOjE2NDYwNTcyNDQsImV4cCI6MTY0NjA2MDg0NH0.NAxgOOzfjtYDwb2-x0iuiCOhEgCWGmY5-YigQw0DCBo").required(),
+  })
+  .label("JwtAuth");