+++ playtime-0.10.0/src/models/mongo/connect.js
@@ -1,6 +1,17 @@
 import * as dotenv from "dotenv";
 import Mongoose from "mongoose";
 
+import { seedData } from "./seed-data.js";
+import * as seeder from "mais-mongoose-seeder";
+
+const seedLib = seeder.default;
+
+async function seed() {
+  const seeder = seedLib(Mongoose);
+  const dbData = await seeder.seed(seedData, { dropDatabase: false, dropCollections: true });
+  console.log(dbData);
+}
+
 export function connectMongo() {
   dotenv.config();
 
@@ -15,7 +26,8 @@ export function connectMongo() {
     console.log("database disconnected");
   });
 
-  db.once("open", function () {
+  db.once("open", function() {
     console.log(`database connected to ${this.name} on ${this.host}`);
+    seed();
   });
 }