+++ playtime-0.10.0/src/server.js
@@ -5,15 +5,15 @@ import Cookie from "@hapi/cookie";
 import dotenv from "dotenv";
 import path from "path";
 import Joi from "joi";
+import jwt from "hapi-auth-jwt2";
 import HapiSwagger from "hapi-swagger";
 import { fileURLToPath } from "url";
 import Handlebars from "handlebars";
 import { webRoutes } from "./web-routes.js";
-import { apiRoutes } from "./api-routes.js";
 import { db } from "./models/db.js";
 import { accountsController } from "./controllers/accounts-controller.js";
-import jwt from "hapi-auth-jwt2";
 import { validate } from "./api/jwt-utils.js";
+import { apiRoutes } from "./api-routes.js";
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
@@ -27,7 +27,7 @@ if (result.error) {
 const swaggerOptions = {
   info: {
     title: "Playtime API",
-    version: "0.1",
+    version: "0.1"
   },
   securityDefinitions: {
     jwt: {
@@ -42,55 +42,54 @@ const swaggerOptions = {
 async function init() {
   const server = Hapi.server({
     port: 3000,
-    host: "localhost",
+    host: "localhost"
   });
 
   await server.register(Inert);
   await server.register(Vision);
   await server.register(Cookie);
   await server.register(jwt);
-  server.validator(Joi);
 
   await server.register([
     Inert,
     Vision,
     {
       plugin: HapiSwagger,
-      options: swaggerOptions,
-    },
+      options: swaggerOptions
+    }
   ]);
+
+  server.validator(Joi);
 
   server.views({
     engines: {
-      hbs: Handlebars,
+      hbs: Handlebars
     },
     relativeTo: __dirname,
     path: "./views",
     layoutPath: "./views/layouts",
     partialsPath: "./views/partials",
     layout: true,
-    isCached: false,
+    isCached: false
   });
 
   server.auth.strategy("session", "cookie", {
     cookie: {
       name: process.env.cookie_name,
       password: process.env.cookie_password,
-      isSecure: false,
+      isSecure: false
     },
     redirectTo: "/",
-    validateFunc: accountsController.validate,
+    validateFunc: accountsController.validate
   });
-
   server.auth.strategy("jwt", "jwt", {
     key: process.env.cookie_password,
     validate: validate,
-    verifyOptions: { algorithms: ["HS256"] },
+    verifyOptions: { algorithms: ["HS256"] }
   });
-
   server.auth.default("session");
 
-  db.init("json");
+  db.init("mongo");
   server.route(webRoutes);
   server.route(apiRoutes);
   await server.start();