+++ playtime-0.10.0/test/api/auth-test.js
@@ -1,27 +1,40 @@
 import { assert } from "chai";
 import { playtimeService } from "./playtime-service.js";
 import { decodeToken } from "../../src/api/jwt-utils.js";
-import { maggie } from "../fixtures.js";
+import { maggie, maggieCredentials } from "../fixtures.js";
 
 suite("Authentication API tests", function() {
 
   setup(async function() {
+    playtimeService.clearAuth()
+    await playtimeService.createUser(maggie);
+    await playtimeService.authenticate(maggieCredentials);
     await playtimeService.deleteAllUsers();
   });
 
   test("authenticate", async function() {
     const returnedUser = await playtimeService.createUser(maggie);
-    const response = await playtimeService.authenticate(maggie);
+    const response = await playtimeService.authenticate(maggieCredentials);
     assert(response.success);
     assert.isDefined(response.token);
   });
 
   test("verify Token", async function() {
     const returnedUser = await playtimeService.createUser(maggie);
-    const response = await playtimeService.authenticate(maggie);
+    const response = await playtimeService.authenticate(maggieCredentials);
 
     const userInfo = decodeToken(response.token);
     assert.equal(userInfo.email, returnedUser.email);
     assert.equal(userInfo.userId, returnedUser._id);
+  });
+
+  test("check Unauthorized", async function() {
+    playtimeService.clearAuth()
+    try {
+      await playtimeService.deleteAllUsers();
+      assert.fail("Route not protected");
+    } catch (error) {
+      assert.equal(error.response.data.statusCode, 401);
+    }
   });
 });