+++ playtime-0.10.0/test/api/playtime-service.js
@@ -1,5 +1,5 @@
 import axios from "axios";
-import { serviceUrl } from "../fixtures.js";
+import { maggie, serviceUrl } from "../fixtures.js";
 
 export const playtimeService = {
   playtimeUrl: serviceUrl,
@@ -15,8 +15,12 @@ export const playtimeService = {
   },
 
   async getAllUsers() {
-    const res = await axios.get(`${this.playtimeUrl}/api/users`);
-    return res.data;
+    try {
+      const res = await axios.get(`${this.playtimeUrl}/api/users`);
+      return res.data;
+    } catch (e) {
+      return null;
+    }
   },
 
   async deleteAllUsers() {
@@ -80,7 +84,7 @@ export const playtimeService = {
     return response.data;
   },
 
-  async clearAuth(user) {
+  async clearAuth() {
     axios.defaults.headers.common["Authorization"] = "";
-  }
+  },
 };