+++ playtime-0.10.0/test/api/tracks-api-test.js
@@ -1,17 +1,21 @@
 import { assert } from "chai";
 import { assertSubset } from "../test-utils.js";
 import { playtimeService } from "./playtime-service.js";
-import { maggie, mozart, testPlaylists, testTracks, concerto } from "../fixtures.js";
+import { maggie, mozart, testPlaylists, testTracks, concerto, maggieCredentials } from "../fixtures.js";
 
 suite("Track API tests", () => {
   let user = null;
   let beethovenSonatas = null;
 
   setup(async () => {
-    await playtimeService.deleteAllPlaylists();
-    await playtimeService.deleteAllUsers();
-    await playtimeService.deleteAllTracks();
+    playtimeService.clearAuth();
     user = await playtimeService.createUser(maggie);
+    await playtimeService.authenticate(maggieCredentials);
+    await playtimeService.deleteAllPlaylists();
+    await playtimeService.deleteAllTracks();
+    await playtimeService.deleteAllUsers();
+    user = await playtimeService.createUser(maggie);
+    await playtimeService.authenticate(maggieCredentials);
     mozart.userid = user._id;
     beethovenSonatas = await playtimeService.createPlaylist(mozart);
   });