+++ playtime-0.10.0/test/api/users-api-test.js
@@ -1,18 +1,23 @@
 import { assert } from "chai";
 import { assertSubset } from "../test-utils.js";
 import { playtimeService } from "./playtime-service.js";
-import { maggie, testUsers } from "../fixtures.js";
+import { maggie, maggieCredentials, testUsers } from "../fixtures.js";
 import { db } from "../../src/models/db.js";
 
 const users = new Array(testUsers.length);
 
 suite("User API tests", () => {
   setup(async () => {
+    playtimeService.clearAuth();
+    await playtimeService.createUser(maggie);
+    await playtimeService.authenticate(maggieCredentials);
     await playtimeService.deleteAllUsers();
     for (let i = 0; i < testUsers.length; i += 1) {
       // eslint-disable-next-line no-await-in-loop
       users[0] = await playtimeService.createUser(testUsers[i]);
     }
+    await playtimeService.createUser(maggie);
+    await playtimeService.authenticate(maggieCredentials);
   });
   teardown(async () => {});
 
@@ -24,10 +29,12 @@ suite("User API tests", () => {
 
   test("delete all user", async () => {
     let returnedUsers = await playtimeService.getAllUsers();
-    assert.equal(returnedUsers.length, 3);
+    assert.equal(returnedUsers.length, 4);
     await playtimeService.deleteAllUsers();
+    await playtimeService.createUser(maggie);
+    await playtimeService.authenticate(maggieCredentials);
     returnedUsers = await playtimeService.getAllUsers();
-    assert.equal(returnedUsers.length, 0);
+    assert.equal(returnedUsers.length, 1);
   });
 
   test("get a user", async () => {
@@ -47,6 +54,8 @@ suite("User API tests", () => {
 
   test("get a user - deleted user", async () => {
     await playtimeService.deleteAllUsers();
+    await playtimeService.createUser(maggie);
+    await playtimeService.authenticate(maggieCredentials);
     try {
       const returnedUser = await playtimeService.getUser(users[0]._id);
       assert.fail("Should not return a response");