+++ playtime-0.9.0/test/api/playlist-api-test.js
@@ -1,7 +1,6 @@
 import { assert } from "chai";
 import { playtimeService } from "./playtime-service.js";
 import { assertSubset } from "../test-utils.js";
-
 import { maggie, mozart, testPlaylists } from "../fixtures.js";
 
 suite("Playlist API tests", () => {
@@ -9,6 +8,9 @@ suite("Playlist API tests", () => {
   let user = null;
 
   setup(async () => {
+    playtimeService.clearAuth();
+    user = await playtimeService.createUser(maggie);
+    await playtimeService.authenticate(maggie);
     await playtimeService.deleteAllPlaylists();
     await playtimeService.deleteAllUsers();
     user = await playtimeService.createUser(maggie);