+++ playtime-0.9.0/test/api/tracks-api-test.js
@@ -8,11 +8,14 @@ suite("Track API tests", () => {
   let beethovenSonatas = null;
 
   setup(async () => {
-    await playtimeService.deleteAllUsers();
+    playtimeService.clearAuth();
     user = await playtimeService.createUser(maggie);
     await playtimeService.authenticate(maggie);
     await playtimeService.deleteAllPlaylists();
     await playtimeService.deleteAllTracks();
+    await playtimeService.deleteAllUsers();
+    user = await playtimeService.createUser(maggie);
+    await playtimeService.authenticate(maggie);
     mozart.userid = user._id;
     beethovenSonatas = await playtimeService.createPlaylist(mozart);
   });